- Old simulator:     - Added several FGCS experiments configuration files.     - Removed dependency with Sqlite3 and Wt.     - Print node properties in CPU statistics.     - CDF optimization replaces sequences with first and last value, not only first.     - cdf program calculates CDF of accepted and rejected tasks by length.     - Rewritten random schedulers, with blind option.     - Database:         - Instance numeration is per node.         - Requests are in a list per instance.         - Request id is int64_t, first bits request id, last bits instance id.         - Last used instance and request are cached. - Sweeper:     - Added a dummy test case.     - sweeper now only generates config files. - Lib:     - Database ints are always int64_t, like request id.     - Use max queue and beta value for QueueBalancingDispatcher

+++ include/Scheduler.hpp
@@ -76,7 +76,7 @@ public:
          * @param ctId The ID of this task relative to its request.
          * @param d TaskDescription with the task requirements.
          */
-        virtual boost::shared_ptr<Task> createTask(CommAddress o, long int reqId, unsigned int ctid, const TaskDescription & d) const = 0;
+        virtual boost::shared_ptr<Task> createTask(CommAddress o, int64_t reqId, unsigned int ctid, const TaskDescription & d) const = 0;
     };
 
     Scheduler(ResourceNode & rn) : ResourceNodeObserver(rn), seqNum(0),
@@ -176,7 +176,7 @@ private:
     int monitorTimer;      ///< Timer to send monitoring info
 
     // Statistics
-    void queueChangedStatistics(unsigned int rid, unsigned int numAccepted, Time queueEnd);
+    void queueChangedStatistics(int64_t rid, unsigned int numAccepted, Time queueEnd);
     unsigned long int tasksExecuted;   ///< Number of executed tasks since the peer started
     Duration timeRunning;              ///< Amount of time not idle
